chore/provisioner: move metrics to a separate module
